% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_own_trade.R
\name{gen_own_trade}
\alias{gen_own_trade}
\title{Generate small own trade values as needed}
\usage{
gen_own_trade(iot, max_replace)
}
\arguments{
\item{iot}{a tibble with a standardized input output table in long format
with the columns \code{origin}, \code{sector}, \code{destination},
\code{use}, \code{flow}. See the package \code{iotr} for more details on this
format.}

\item{max_replace}{maximum value to replace zero own trade flows with.}
}
\value{
Returns the data frame \code{iot} with respective zero own trade
flows removed.
}
\description{
Finds combinations of destination-sector-use in a standardized IO table that
are 0 where \code{destination == origin} but different from 0 for at least
one \code{destination != origin} and replaces them with either the smallest
value found for one of the \code{origin != destination} cases or
\code{max_replace}, whichever is smaller.
}
\references{
This manipulation of the original IO table is useful for
specific simulations that feature infinite trade barrieres. See for example
Eppinger, Peter and Felbermayr, Gabriel J. and Krebs, Oliver and
Kukharskyy, Bohdan, COVID-19 Shocking Global Value Chains (2020). CESifo
Working Paper No. 8572, https://ssrn.com/abstract=3702124
}
