% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wiot.R
\name{get_wiot}
\alias{get_wiot}
\title{Download IO tables from the WIOD}
\usage{
get_wiot(
  cache_dir = NULL,
  years = 2000:2014,
  url = paste0("http://www.wiod.org/protected3/", "data16/wiot_ROW/wiot_r_Nov16.zip")
)
}
\arguments{
\item{cache_dir}{path to cache the WIOD data. If the full content of the WIOD
data zip file is found under this location it will not be downloaded again.}

\item{years}{which year(s) to load. Must be a single value or vector of
values between 2000 and 2014}

\item{url}{url to download the WIOD from. This should usually not be changed.}
}
\value{
Returns one \code{data.frame} combining all wiots from the WIOD's
2016 release
}
\description{
Function to download the set of world input output tables (wiot) made
available by the World Input Output Database's (WIOD) 2016 release.
}
\examples{
library(iotr)
library(dplyr)

wiot <- get_wiot(cache_dir = tempdir(), years = 2014)
wiot_long <- wiot2long(wiot)

# get wiot into standard long IO-table format
iot <- wiot_long \%>\%
  filter(Country != "TOT") \%>\%
  select(origin = Country, sector = RNr, destination, use, flow)

iot <- remove_negative_vad(iot, category_to_scale = 57)

iot <- remove_dynamic_categories(iot,
                                 dynamic_categories = c(60, 61),
                                 category_to_scale = 57)
}
\references{
Timmer, M. P., Dietzenbacher, E., Los, B., Stehrer, R. and de
Vries, G. J. (2015), "An Illustrated User Guide to the World Input–Output
Database: the Case of Global Automotive Production", Review of International
Economics., 23: 575–605, www.wiod.org
}
