% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{io_correspondences}
\alias{io_correspondences}
\title{Sector and Country/Location Correspondences}
\format{
A list with each element representing one of the data sets in the
package. Each element is itself a list containing
\describe{
\item{wiot}{Sectors and Locations in the World Input Output Database}
\itemize{
\item{locations:}
\describe{
\item{location_id}{id for each location, with alphabetically ordered iso3}
\item{wiot_location_id}{same as above but with ROW assigned the last id}
\item{wiot_location_iso3}{country iso3 as provided by WIOD}
\item{wiot_location}{full name as provided by WIOD}
}
}
}
}
\usage{
io_correspondences
}
\description{
A list that, for each IO data set in the packages, provides a sector and a
country/locations correspondence.
}
\keyword{datasets}
