% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dl_eora26.R
\name{io_dl_eora26}
\alias{io_dl_eora26}
\title{Download IO tables from the Eora26 data set}
\usage{
io_dl_eora26(
  years,
  cache_dir = NULL,
  email = NULL,
  password = NULL,
  quiet = FALSE
)
}
\arguments{
\item{years}{which year(s) to load. Must be a single value or vector of
values between 1990 and the latest year available with your account.}

\item{cache_dir}{path to cache the Eora26 data. Downloads will overwrite
files of the same name in this location.}

\item{email}{the email adress of the account}

\item{password}{the password of the account}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Path to the directory where data was saved, if given, this is
identical to \code{cache_dir}
}
\description{
Function to download the multiregional input output (MRIO) tables of the
Eora26. Should usually not be called directly but through
\code{io_load_eora26()}. Note this tool simply automates the download process
but you need to obtain an account from https://worldmrio.com and agree to
their usage and licensing terms to download and use this data.
}
\references{
Lenzen M, Kanemoto K; Moran D, and Geschke A (2012) Mapping the
structure of the world economy. Environmental Science & Technology 46(15)
pp 8374–8381.

Lenzen, M., Moran, D., Kanemoto, K., Geschke, A. (2013) Building Eora: A
Global Multi-regional Input-Output Database at High Country and Sector
Resolution. Economic Systems Research, 25:1, 20-49.
}
