% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dl_icio_v2.R
\name{io_dl_icio_v2}
\alias{io_dl_icio_v2}
\title{Download IO tables from the OECD ICIO v2}
\usage{
io_dl_icio_v2(
  year_blk,
  cache_dir = NULL,
  url = list(base_url = "https://stats.oecd.org/wbos/fileview2.aspx?IDFile=", `1995-2000`
    = "d26ad811-5b58-4f0c-a4e3-06a1469e475c", `2001-2005` =
    "7cb93dae-e491-4cfd-ac67-889eb7016a4a", `2006-2010` =
    "ea165bfb-3a85-4e0a-afee-6ba8e6c16052", `2011-2015` =
    "1f791bc6-befb-45c5-8b34-668d08a1702a", `2016-2020` =
    "d1ab2315-298c-4e93-9a81-c6f2273139fe"),
  quiet = FALSE
)
}
\arguments{
\item{year_blk}{a single integer between 1 and 5, denoting which year block
(1995-2000, 2001-2005, 2006-2010, 2011-2015, 2016-2020) to download.}

\item{cache_dir}{path to cache the ICIO data. If the full content of the ICIO
data zip file is found under this location it will not be downloaded again.}

\item{url}{list of urls to download the ICIO zip files from. This should
usually not be changed.}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Path to the directory where data was saved, if given, this is
identical to \code{cache_dir}
}
\description{
Function to download the set of inter country input output (icio) tables made
available by the OECD's 2023 release. Should usually not be called directly
but through \code{io_load_icio()}.
}
\references{
OOECD. (2023) OECD Inter-Country Input-Output Tables,
http://oe.cd/icio
}
