% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dl_wiot.R
\name{io_dl_wiot}
\alias{io_dl_wiot}
\title{Download IO tables from the WIOD}
\usage{
io_dl_wiot(
  cache_dir = NULL,
  url = paste0("https://dataverse.nl/api/access/datafile/199101"),
  quiet = FALSE
)
}
\arguments{
\item{cache_dir}{path to cache the WIOD data. If the full content of the WIOD
data zip file is found under this location it will not be downloaded again.}

\item{url}{url to download the WIOD from. This should usually not be changed.}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Path to the directory where data was saved, if given, this is
identical to \code{cache_dir}
}
\description{
Function to download the set of world input output tables (wiot) made
available by the World Input Output Database's (WIOD) 2016 release.
}
\examples{
#'\dontrun{

library(iotr)

# set cache_dir for WIOD to avoid long download times on every run !!!
cache_dir <- NULL

wiot_raw <- io_load_wiot(cache_dir, years = 2000:2014)

wiot_long <- io_tidy_wiot(wiot_raw)

# get wiot into standard long IO-table format
iot <- dplyr::filter(wiot_long, Country != "TOT", Year == "2014")
iot <-
  dplyr::select(iot,
                origin = Country,
                sector = RNr,
                destination,
                use,
                flow)

iot <- io_rm_negative_vad(iot, category_to_scale = 57)

iot <-
  io_rm_dynamics(iot,
                 dynamic_categories = c(60, 61),
                 category_to_scale = 57)

iot <- io_gen_own_trade(iot, max_replace = 1e-6)

#' }
}
\references{
Timmer, M. P., Dietzenbacher, E., Los, B., Stehrer, R. and de
Vries, G. J. (2015), "An Illustrated User Guide to the World Input–Output
Database: the Case of Global Automotive Production", Review of International
Economics., 23: 575–605, www.wiod.org
}
