% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_load_eora26.R
\name{io_load_eora26}
\alias{io_load_eora26}
\title{Load IO tables from the EORA26 MRIO}
\usage{
io_load_eora26(
  years,
  cache_dir = NULL,
  email = NULL,
  password = NULL,
  version = NULL,
  quiet = FALSE
)
}
\arguments{
\item{years}{which year(s) to load. Must be a single value or vector of
values between 1990 and the latest year available with your account.}

\item{cache_dir}{path to cache the Eora26 data. Downloads will overwrite
files of the same name in this location.}

\item{email}{the email adress of the account}

\item{password}{the password of the account}

\item{version}{the EORA26 version to load. If NULL the most recent version in
the cache will be used if any is available, otherwise, or if "current" the
most recent version will be checked from the website and if necessary
downloaded}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Returns a list combining for each requested year, the T, VA and FD
components from the requested version of the EORA26 as well as as a list
element with the corresponding labels.
}
\description{
Function to load and, if not cached locally, download the set of
multiregional input output (MRIO) tables of the EORA. Note that this tool
simply automates the download process but you need to obtain an account from
https://worldmrio.com and agree to their usage and licensing terms to
download and use this data.
}
\examples{
\dontrun{

  library(iotr)

  # set cache_dir for EORA to avoid (very) long download times on every run !!!
  cache_dir <- NULL

  eora26_raw <- io_load_icio_v2(cache_dir, years = 2016:2017)

  eora26 <- io_tidy_eora26(eora26_raw)
}
}
\references{
Lenzen M, Kanemoto K; Moran D, and Geschke A (2012) Mapping the
structure of the world economy. Environmental Science & Technology 46(15)
pp 8374–8381.

Lenzen, M., Moran, D., Kanemoto, K., Geschke, A. (2013) Building Eora: A
Global Multi-regional Input-Output Database at High Country and Sector
Resolution. Economic Systems Research, 25:1, 20-49.
}
