% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_load_icio_v2.R
\name{io_load_icio_v2}
\alias{io_load_icio_v2}
\title{Download IO tables from the OECD ICIO v2}
\usage{
io_load_icio_v2(cache_dir = NULL, years = 1995:2020, quiet = FALSE)
}
\arguments{
\item{cache_dir}{path to cache the ICIO data. If the full content of the ICIO
data zip file is found under this location it will not be downloaded again.}

\item{years}{which year(s) to load. Must be a single value or vector of
values between 1995 and 2020. Can be used to reduce memory needs and speed
up loading.}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Returns one \code{data.frame} combining all years and requested
components from the ICIOs from the OECD's 2023 release.
}
\description{
Function to download the set of inter country input output (icio) tables made
available by the OECD's 2023 release.
}
\examples{
\dontrun{

  library(iotr)

  # set cache_dir for ICIO to avoid (very) long download times on every run !!!
  cache_dir <- NULL

  icio_v2_raw <- io_load_icio_v2(cache_dir, years = 2010:2012)

  icio_v2 <- io_tidy_icio_v2(icio_v2_raw)
}
}
\references{
OECD. (2023) OECD Inter-Country Input-Output Tables,
http://oe.cd/icio
}
