% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_plot_icio_map.R
\name{io_plot_icio_map}
\alias{io_plot_icio_map}
\title{Visualize data for ICIO countries}
\usage{
io_plot_icio_map(
  plot_data,
  row_label = FALSE,
  percent = FALSE,
  suffix = NULL,
  digits = 2,
  reverse_colors = FALSE
)
}
\arguments{
\item{plot_data}{dataframe/tibble containint a column "country" with iso3
codes from the ICIO and another column holding the data to visualize on a
map}

\item{row_label}{if TRUE a label with the value of 'ROW' in '\code{plot_data}'
will be placed in the ocean west of Australia}

\item{percent}{multiplies label values by 100, TRUE/FALSE}

\item{suffix}{character suffix to add to labels, e.g. "\%"}

\item{digits}{number of digits to round labels to}

\item{reverse_colors}{if TRUE color scale will be reversed}
}
\value{
Return a plot of all ICIO countries filled with the appropriate color
}
\description{
Plots a concise map of all countries from the OECD ICIO Database, split into
three groups (broadly Asia, Americas, Europe), to obtain a larger
visualization for individual countries. Countries will be colored according
to the dataset passed to the function. Due to its geographic location the
value for south africa is not visualized but added as a label (ZAF).
}
