% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_plot_wiot_map.R
\name{io_plot_wiot_map}
\alias{io_plot_wiot_map}
\title{Visualize data for WIOD countries}
\usage{
io_plot_wiot_map(plot_data, row_label = FALSE)
}
\arguments{
\item{plot_data}{dataframe/tibble containint a column "country" with iso3
codes from the WIOD and another column holding the data to visualize on a
map}

\item{row_label}{if TRUE a label with the value of 'ROW' in '\code{plot_data}'
will be placed in the ocean west of Australia}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Return a plot of all WIOD countries filled with the appropriate color
}
\description{
Plots a concise map of all countries from the World Input Output Database's
(WIOD) 2016 release, spit into three groups (broadly Asia, Americas, Europe),
to obtain a larger visualization for individual countries. Countries will be
colored according to the dataset passed to the function.
}
