% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_set_cache_dir.R
\name{io_set_cache_dir}
\alias{io_set_cache_dir}
\title{Set paths to cache large IO downloads}
\usage{
io_set_cache_dir(cache_dir = NULL, io_database = NULL)
}
\arguments{
\item{cache_dir}{If set to a character this will replace the current cache
path stored in the iotr configs.}

\item{io_database}{For which database should the cache directory be set.
Current options are 'wiod'  or 'icio'.}
}
\value{
Returns path to the respective cache directory.
}
\description{
Set paths to cache large IO downloads
}
\details{
Several \verb{io_dl_*} and \verb{io_load_*} functions ask for a cache
directory to cache several large IO data files in order to avoid repeated
downloads. This function is a helper to (re)set the folder paths where to
keep these files. The given path is stored in a config file (at the user
configs location obtained from package rappdirs). Manually call this funtion
with a new path to reset it.
}
