% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tidy_icio.R
\name{io_tidy_icio}
\alias{io_tidy_icio}
\title{Tidy data from OECD's ICIO}
\usage{
io_tidy_icio(icio, quiet = FALSE)
}
\arguments{
\item{icio}{list containing ICIO components}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Returns one \code{tibble} combining all components from \code{icio}.
}
\description{
Function to turn a list of OECD ICIO components, usually obtained from a call
to \code{io_load_icio()}, into a tibble, i.e. a 'tidy' data frame.
}
\examples{
library(iotr)

# set cache_dir for ICIO to avoid (very) long download times on every run !!!
cache_dir <- NULL

icio_raw <- io_load_icio(cache_dir,
                         years = 2010:2012,
                         components = c("Z", "CONS", "GFCF", "INVNT", "NONRES"))

icio <- io_tidy_icio(icio_raw)
}
\references{
OECD (2021), OECD Inter-Country Input-Output Database,
http://oe.cd/icio
}
