% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tidy_icio_v2.R
\name{io_tidy_icio_v2}
\alias{io_tidy_icio_v2}
\title{Tidy data from OECD's ICIO v2}
\usage{
io_tidy_icio_v2(icio, quiet = FALSE)
}
\arguments{
\item{icio}{list containing ICIO years from the 2023 release}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Returns one \code{tibble} combining all components from \code{icio}.
}
\description{
Function to turn a list of OECD ICIO components, usually obtained from a call
to \code{io_load_icio_v2()}, into a tibble, i.e. a 'tidy' data frame.
}
\examples{
\dontrun{

  library(iotr)

  # set cache_dir for ICIO to avoid (very) long download times on every run !!!
  cache_dir <- NULL

  icio_v2_raw <- io_load_icio_v2(cache_dir, years = 2010:2012)

  icio_v2 <- io_tidy_icio_v2(icio_v2_raw)
}
}
\references{
OECD. (2023) OECD Inter-Country Input-Output Tables,
http://oe.cd/icio
}
