% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotr-package.R
\docType{package}
\name{iotr-package}
\alias{iotr}
\alias{iotr-package}
\title{iotr: Obtain, Tidy and Modify Input-Output Tables For Quantitative Economics}
\description{
Package for working with Input Output Tables in quantitativ economics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/okrebs/iotr}
  \item Report bugs at \url{https://github.com/okrebs/iotr/issues}
}

}
\author{
\strong{Maintainer}: Oliver Krebs \email{oliver.b.krebs@gmail.cm} (\href{https://orcid.org/0000-0002-6217-5880}{ORCID})

}
\keyword{internal}
