% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_dynamics.R
\name{rm_dynamics}
\alias{rm_dynamics}
\title{Remove Dynamic Consumption Categories from IO Table}
\usage{
rm_dynamics(iot, dynamic_categories, category_to_scale)
}
\arguments{
\item{iot}{An input-output table in long format with the columns,
\code{origin}, \code{sector}, \code{destination}, \code{use} and
\code{flow}}

\item{dynamic_categories}{a vector of dynamic categories in final use to be
removed}

\item{category_to_scale}{the use category (as integer) to which to add the
positive components of removed categories.}
}
\value{
Returns an IO table in long format as a \code{tibble} with the
columns \code{origin}, \code{sector}, \code{destination}, \code{use} and
\code{flow}
}
\description{
Remove dynamic consumption categories, e.g. inventory or
gross-fixed-capital-formation from an input-output (IO) table by
recalculating the IO table based on its leontief inverse and implied new
demand.
}
\details{
Static CGE trade models often can/do not capture dynamic components
such as changes in inventory or gross-fixed-capital-formation. Therefore
this function allows to recalculate the IO table following the suggestions
in Costinot and Rodríguez-Clare (2014). Specifically, negative changes in
dynamic components of final consumption are assumed to have been produced
in the current period, and positive components are added to a different
category that can be specified using the \code{category_to_scale} argument
(usually this will be final consumption by households). To account for the
intermediate goods that would have been necessary to produce the negative
components in the current period the approach relies on a leontief inverse
to calculate intermediate use based upon the implied new demand.
}
\examples{
library(iotr)
library(dplyr)

wiot <- get_wiot(cache_dir = tempdir(), years = 2014)
wiot_long <- wiot2long(wiot)

# get wiot into standard long IO-table format
iot <- wiot_long \%>\%
  filter(Country != "TOT") \%>\%
  select(origin = Country, sector = RNr, destination, use, flow)

iot <- rm_negative_vad(iot, category_to_scale = 57)

iot <- rm_dynamics(iot, dynamic_categories = c(60, 61), category_to_scale = 57)

iot <- gen_own_trade(iot, max_replace = 1e-6)
}
