% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiot2long.R
\name{wiot2long}
\alias{wiot2long}
\title{Convert WIOT table to long format}
\usage{
wiot2long(wiot)
}
\arguments{
\item{wiot}{A \code{data.frame} containing a WIOT for one or many years.}
}
\value{
Returns one \code{tibble} with the WIOT tables converted to a long
format
}
\description{
Converts WIOT table(s) from the 2016 release of the World Input Output
Database into a long format.
}
\examples{
library(iotr)
library(dplyr)

wiot <- get_wiot(cache_dir = tempdir(), years = 2014)
wiot_long <- wiot2long(wiot)

# get wiot into standard long IO-table format
iot <- wiot_long \%>\%
  filter(Country != "TOT") \%>\%
  select(origin = Country, sector = RNr, destination, use, flow)

iot <- remove_negative_vad(iot, category_to_scale = 57)

iot <- remove_dynamic_categories(iot,
                                 dynamic_categories = c(60, 61),
                                 category_to_scale = 57)
}
\seealso{
\code{\link{get_wiot}} for downloading WIOT tables
}
