% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dl_wiot.R
\name{io_dl_wiot}
\alias{io_dl_wiot}
\title{Download IO tables from the WIOD}
\usage{
io_dl_wiot(
  cache_dir = NULL,
  url = paste0("https://dataverse.nl/api/access/datafile/199101"),
  quiet = FALSE
)
}
\arguments{
\item{cache_dir}{path to cache the WIOD data. If the full content of the WIOD
data zip file is found under this location it will not be downloaded again.}

\item{url}{url to download the WIOD from. This should usually not be changed.}

\item{quiet}{if TRUE will try to avoid printing messages}
}
\value{
Path to the directory where data was saved, if given, this is
identical to \code{cache_dir}
}
\description{
Function to download the set of world input output tables (wiot) made
available by the World Input Output Database's (WIOD) 2016 release. Should
usually not be called directly but through \code{io_load_wiot()}.
}
\references{
Timmer, M. P., Dietzenbacher, E., Los, B., Stehrer, R. and de
Vries, G. J. (2015), "An Illustrated User Guide to the World Input–Output
Database: the Case of Global Automotive Production", Review of International
Economics., 23: 575–605, www.wiod.org
}
